fix: Improve CUDA Graph stability and fix preview window rendering
- Add BGRA to RGBA conversion kernel for correct color handling
- Fix SimpleCudaMat copyFrom to handle different step sizes properly
- Improve CUDA Graph capture with retry logic and validation
- Add comprehensive error handling and fallback mechanisms
- Wait 30 frames before attempting CUDA Graph capture
- Validate graph nodes and test launch before enabling
- Auto-disable after 5 failed attempts
- Remove debug logging from preview window
- Fix compiler warning in image_io.cpp

+++ needaimbot/utils/image_io.cpp
@@ -22,7 +22,7 @@ bool saveImage(const SimpleMat& image, const std::string& filename, int quality)
     int width = image.cols();
     int height = image.rows();
     int channels = image.channels();
-    int stride = image.step();
+    int stride = static_cast<int>(image.step());
     
     int result = 0;
